refactor: 1주차 PR review 반영
