+++ vending-machine/src/components/vendingMachine/showWindowItemBox/itemBox.styled.js
