+++ vending-machine/src/components/vendingMachine/userWindowLogMonitor/logMonitor.js
