+++ vending-machine/src/components/vendingMachine/userWindowLogMonitor/logMonitor.styled.js
